Re-allow Xom to gift useless items
Well, sort of. He already could through another codepath, and he
couldn't actually previously gift useless armour through this
path... maybe this code shouldn't exist. I'm undecided.

+++ crawl-ref/source/acquire.cc
@@ -44,6 +44,7 @@ static armour_type _acquirement_shield_type();
 static armour_type _acquirement_body_armour(bool);
 static armour_type _acquirement_mundane_armour(bool);
 static armour_type _acquirement_hide_armour(bool);
+static armour_type _useless_armour_type();
 
 
 /**
@@ -320,6 +321,76 @@ static armour_type _acquirement_mundane_armour(bool divine)
     }
 
     return sub_type;
+}
+
+/**
+ * Choose a random type of armour that the player cannot wear, for Xom to spite
+ * the player with.
+ *
+ * @return  A random useless armour_type.
+ */
+static armour_type _useless_armour_type()
+{
+    vector<pair<equipment_type, int>> weights = {
+        { EQ_BODY_ARMOUR, 1 }, { EQ_SHIELD, 1 }, { EQ_CLOAK, 1 },
+        { EQ_HELMET, 1 }, { EQ_GLOVES, 1 }, { EQ_BOOTS, 1 },
+    };
+
+    // everyone has some kind of boot-slot item they can't wear, regardless
+    // of what you_can_wear() claims
+    for (auto &weight : weights)
+        if (you_can_wear(weight.first) == MB_TRUE && weight.first != EQ_BOOTS)
+            weight.second = 0;
+
+    const equipment_type* slot_ptr = random_choose_weighted(weights);
+    const equipment_type slot = slot_ptr ? *slot_ptr : EQ_BOOTS;
+
+    switch (slot)
+    {
+        case EQ_BOOTS:
+            if (you_can_wear(EQ_BOOTS))
+                return coinflip() ? ARM_CENTAUR_BARDING : ARM_NAGA_BARDING;
+            return ARM_BOOTS;
+        case EQ_GLOVES:
+            return ARM_GLOVES;
+        case EQ_HELMET:
+            if (you_can_wear(EQ_HELMET))
+                return ARM_HELMET;
+            return coinflip() ? ARM_HELMET : ARM_HAT;
+        case EQ_CLOAK:
+            return ARM_CLOAK;
+        case EQ_SHIELD:
+        {
+            vector<pair<armour_type, int>> shield_weights = {
+                { ARM_BUCKLER,       1 },
+                { ARM_SHIELD,        1 },
+                { ARM_LARGE_SHIELD,  1 },
+            };
+
+            // XXX: export this idiom ^ v
+            for (auto &weight : shield_weights)
+                if (check_armour_size(weight.first, you.body_size()))
+                    weight.second = 0;
+
+            const armour_type* shield_type
+                = random_choose_weighted(shield_weights);
+            ASSERT(shield_type);
+            return *shield_type;
+        }
+        case EQ_BODY_ARMOUR:
+            // only the rarest & most precious of unwearable armours for Xom
+            if (you_can_wear(EQ_BODY_ARMOUR))
+                return ARM_CRYSTAL_PLATE_ARMOUR;
+            // arbitrary selection of [unwearable] dragon armours
+            return random_choose(ARM_FIRE_DRAGON_ARMOUR,
+                                 ARM_ICE_DRAGON_ARMOUR,
+                                 ARM_PEARL_DRAGON_ARMOUR,
+                                 ARM_GOLD_DRAGON_ARMOUR,
+                                 ARM_SHADOW_DRAGON_ARMOUR,
+                                 ARM_STORM_DRAGON_ARMOUR);
+        default:
+            die("Unknown slot type selected for Xom bad-armour-acq!");
+    }
 }
 
 static armour_type _pick_unseen_armour()
@@ -1136,9 +1207,16 @@ int acquirement_create_item(object_class_type class_wanted,
 #define MAX_ACQ_TRIES 40
     for (int item_tries = 0; item_tries < MAX_ACQ_TRIES; item_tries++)
     {
-        // This may clobber class_wanted (e.g. staves/rods, or vampire food)
-        int type_wanted = _find_acquirement_subtype(class_wanted, quant,
+        int type_wanted = -1;
+        if (agent == GOD_XOM && class_wanted == OBJ_ARMOUR && one_chance_in(20))
+            type_wanted = _useless_armour_type();
+        else
+        {
+            // This may clobber class_wanted (e.g. staves/rods, or vampire food)
+            type_wanted = _find_acquirement_subtype(class_wanted, quant,
                                                     divine, agent);
+        }
+        ASSERT(type_wanted != -1);
 
         // Don't generate randart books in items(), we do that
         // ourselves.
@@ -1219,15 +1297,18 @@ int acquirement_create_item(object_class_type class_wanted,
                     // unfilled slots, we might want to roll again.
                     destroy_item(thing_created, true);
                     thing_created = NON_ITEM;
+                    if (!quiet)
+                        dprf("Destroying plain item!");
                     continue;
                 }
             }
         }
 
-        if (acq_item.base_type == OBJ_WEAPONS
-               && !can_wield(&acq_item, false, true)
-            || acq_item.base_type == OBJ_ARMOUR
-               && !can_wear_armour(acq_item, false, true))
+        if (agent != GOD_XOM
+             && (acq_item.base_type == OBJ_WEAPONS
+                   && !can_wield(&acq_item, false, true)
+                 || acq_item.base_type == OBJ_ARMOUR
+                   && !can_wear_armour(acq_item, false, true)))
         {
             destroy_item(thing_created, true);
             thing_created = NON_ITEM;
